Share servo glue
